# src/main/java/uce/edu/unidad3/pw_u3_p4_pa_cm/controller/EstudianteControllerRestFul.java
package uce.edu.unidad3.pw_u3_p4_pa_cm.controller;

import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.methodOn;

import java.util.List;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.hateoas.Link;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.CrossOrigin;
import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import io.swagger.v3.oas.annotations.tags.Tag;
import uce.edu.unidad3.pw_u3_p4_pa_cm.model.Estudiante;
import uce.edu.unidad3.pw_u3_p4_pa_cm.service.IEstudianteService;
import uce.edu.unidad3.pw_u3_p4_pa_cm.service.IMateriaService;
import uce.edu.unidad3.pw_u3_p4_pa_cm.service.to.EstudianteTO;
import uce.edu.unidad3.pw_u3_p4_pa_cm.service.to.MateriaTO;

@RestController
@Tag(name = "Usuarios", description = "Operaciones con usuarios")
@CrossOrigin("http://localhost:3000/")
@RequestMapping(path = "/estudiantes")
public class EstudianteControllerRestFul {

    @Autowired
    private IEstudianteService estudianteService;

    @Autowired
    private IMateriaService materiaService;

    @GetMapping(path = "/{cedula}", produces = "application/json")
    public Estudiante consultarPorCedula(@PathVariable String cedula) {
        return this.estudianteService.consultarPorCedula(cedula);
    }

    @PostMapping(consumes = "application/json")
    public void ingresarEstudiante(@RequestBody Estudiante estu) {
        this.estudianteService.insertarEstudianteService(estu);
    }

    @PutMapping(path = "/{id}")
    public void actualizarEstudiante(@RequestBody Estudiante estu, @PathVariable Integer id) {
        estu.setId(id);
        this.estudianteService.actualizarEstudianteService(estu);
    }

    @PatchMapping(path = "/{cedula}")
    public void actualizarParcialEstudiante(@RequestBody Estudiante estu, @PathVariable String cedula) {
        Estudiante estu1 = this.estudianteService.consultarPorCedula(cedula);
        if (estu.getNombre() == null) {
        }
        estu1.setCedula(estu.getCedula());
        this.estudianteService.actualizarEstudianteService(estu1);
    }

    @DeleteMapping(path = "/{cedula}")
    public void borrarCedula(@PathVariable String cedula) {
        this.estudianteService.eliminarEstudianteService(cedula);
    }
    
    @GetMapping ( produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<List<EstudianteTO>> buscarTodosHATEOAS() {
        List<EstudianteTO> lista = this.estudianteService.buscarTodosH();
        for (EstudianteTO estudianteTO : lista) {
            Link myLink = linkTo(methodOn(EstudianteControllerRestFul.class).buscarPorEstudiante(estudianteTO.getCedula())).withRel("materias");
            estudianteTO.add(myLink);
        }
        return new ResponseEntity<>(lista, null, 200);
    }

    @GetMapping(path = "/{cedula}/materias", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<List<MateriaTO>> buscarPorEstudiante(@PathVariable String cedula) {
        List<MateriaTO> lista = this.materiaService.buscarPorCedulaEstuService(cedula);
        for (MateriaTO materiaTO : lista) {
            Link myLink = linkTo(methodOn(MateriaController.class).consultarMateria(materiaTO.getId())).withSelfRel();
            materiaTO.add(myLink);
        }
        return new ResponseEntity<List<MateriaTO>>(lista, null, 200);
    }

}