# src/main/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: API de estudiantes
  description: Operaciones para gestionar estudiantes
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Servidor local
paths:
  /estudiantes:
    post:
      tags:
      - Estudiantes
      summary: Ingresar nuevo estudiante
      operationId: ingresarEstudiante
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Estudiante'
        required: true
      responses:
        200:
          description: Estudiante ingresado correctamente
        400:
          description: Solicitud inválida
        500:
          description: Error interno del servidor
    get:
      tags:
      - Estudiantes  
      summary: Obtener lista de estudiantes
      operationId: buscarTodos
      responses:
        200:
          description: Lista de estudiantes
          content:
            application/json:    
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EstudianteTO' 
  /estudiantes/{cedula}:
    get:
      tags:
      - Estudiantes
      summary: Obtener un estudiante por cédula
      operationId: consultarPorCedula
      parameters:
      - name: cedula
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Estudiante encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Estudiante'
        404:
          description: Estudiante no encontrado
    delete:
      tags:
      - Estudiantes
      summary: Eliminar un estudiante
      operationId: borrarCedula
      parameters:
      - name: cedula
        in: path
        required: true
        schema:
          type: string
      responses:
        204:
          description: Estudiante eliminado
        404:
          description: Estudiante no encontrado
    patch:
      tags:
      - Estudiantes
      summary: Actualizar parcialmente un estudiante
      operationId: actualizarParcialEstudiante
      parameters:
      - name: cedula
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Estudiante'
      responses:
        200:
          description: Estudiante actualizado
        404:
          description: Estudiante no encontrado
  /estudiantes/{id}:  
    put:
      tags:
      - Estudiantes
      summary: Actualizar estudiante
      operationId: actualizarEstudiante
      parameters:
      - name: id 
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Estudiante'
      responses:
        200:
          description: Estudiante actualizado
        404:
          description: Estudiante no encontrado
  /estudiantes/{cedula}/materias:
    get:
      tags:
      - Materias
      summary: Obtener materias de un estudiante
      operationId: buscarPorEstudiante
      parameters:
      - name: cedula
        in: path
        required: true
        schema:
          type: string
      responses:
        200:
          description: Lista de materias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MateriaTO'

components:
  schemas:
    Estudiante:
      type: object
      properties:
        id:
          type: integer
          format: int32
        cedula:
          type: string
        nombre:
          type: string
        apellido:
          type: string
    EstudianteTO:
      type: object
      properties:
        cedula:
          type: string 
        nombreCompleto:
          type: string
        materias:
          type: array
          items:
            $ref: '#/components/schemas/MateriaTO'
    MateriaTO:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nombre:
          type: string
        creditos:
          type: integer
          format: int32